+++ usage.py
@@ -4,6 +4,7 @@ import dash
 import pandas as pd
 import plotly.express as px
 from dash_molstar.utils import molstar_helper
+from dash_molstar.utils.representations import Representation
 import dash_bootstrap_components as dbc
 
 app = Dash(__name__, assets_folder='bootstrap')
@@ -14,6 +15,7 @@ app.layout = html.Div([
             html.Button(id='load_protein', children="Load Protein", className="btn btn-primary", style={'padding': '5px', 'margin': '5px'}),
             html.Button(id='load_protein_rep', children="Load Protein With Component", className="btn btn-primary", style={'padding': '5px', 'margin': '5px'}),
             html.Button(id='load_protein_url', children="Load Protein From Internet", className="btn btn-primary", style={'padding': '5px', 'margin': '5px'}),
+            html.Button(id='load_traj', children="Load Trajectroy", className="btn btn-primary", style={'padding': '5px', 'margin': '5px'}),
         ], className='mx-auto')
     ]),
     dbc.Row([
@@ -51,22 +53,43 @@ app.layout = html.Div([
           prevent_initial_call=True)
 def load_protein(yes, yess, yesss):
     if not ctx.triggered_id == 'load_protein_url':
-        CDRs = [
-            # CDRs on heavy chain
-            molstar_helper.get_targets(chain="H", residue=[24,25,26,27,28,29,30,31,49,50,51,52,53,54,55,56,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109]),
-            # CDRs on light chain
-            molstar_helper.get_targets(chain="L", residue=[24,25,26,27,45,46,47,84,85,86,87,88])
-        ]
-        cdrs = molstar_helper.create_component("CDRs", CDRs, 'orientation')
-        ag = molstar_helper.create_component("Antigen", molstar_helper.get_targets(chain="G"), 'molecular-surface')
         component = None
         if ctx.triggered_id == 'load_protein_rep':
+            # select the targets
+            CDRs = [
+                # CDRs on heavy chain
+                molstar_helper.get_targets(chain="H", residue=[24,25,26,27,28,29,30,31,49,50,51,52,53,54,55,56,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109]),
+                # CDRs on light chain
+                molstar_helper.get_targets(chain="L", residue=[24,25,26,27,45,46,47,84,85,86,87,88])
+            ]
+            # create the components
+            cdr_rep = Representation(type='orientation')
+            cdr_rep.set_type_params({'alpha': 0.5})
+            # cdr_rep = Representation(type='cartoon', color='element-symbol', size='physical')
+            # cdr_rep.set_color_params({
+            #         "carbonColor": cdr_rep.np("chain-id", {"palette": cdr_rep.np("generate", {"hue": [55, 180]}), 
+            #                                                "asymId": "label"})
+            #         })
+            # cdr_rep.set_size_params({'scale': 1.3})
+            # cdr_rep = Representation.from_config("/Volumes/ToshibaTBT/simon/Downloads/test.json")
+            cdrs = molstar_helper.create_component("CDRs", CDRs, cdr_rep)
+            ag_rep = Representation('molecular-surface')
+            ag_rep.set_type_params({'xrayShaded': True})
+            ag = molstar_helper.create_component("Antigen", molstar_helper.get_targets(chain="G"), ag_rep)
             component=[ag,cdrs]
-        data = molstar_helper.parse_molecule('3u7y.pdb',component=component)
+        data = molstar_helper.parse_molecule('3u7y.pdb', component=component)
     else:
         data = molstar_helper.parse_url('https://files.rcsb.org/download/3U7Y.pdb')
     return data
 
+@callback(Output('viewer', 'data', allow_duplicate=True), 
+          Input('load_traj', 'n_clicks'),
+          prevent_initial_call=True)
+def load_trajectory(yes):
+    topo = molstar_helper.parse_url('/assets/traj.pdb')
+    coord = molstar_helper.parse_url('/assets/traj.dcd')
+    data = molstar_helper.get_trajectory(topo, coord)
+    return data
 
 @callback(Output('viewer', 'selection'),
           Output('viewer', 'focus'),
@@ -103,4 +126,4 @@ def mouse_event(hoverData, clickData, onclick):
 
 
 if __name__ == '__main__':
-    app.run_server(debug=True, )
+    app.run_server(debug=True,)