+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # dash-molstar Changelog
 
+## [1.2.0] - 2024-09-23
+### Added
+- Loading trajectories now supported
+- Loading structures without default representation
+- Fully customizable component representations
+- Import and export representation configurations as JSON file
+- Showing assembly when loading structures
+- Drawing spheres on the stage
+- Allowing transparency on shapes
+
 ## [1.1.2] - 2024-02-22
 ### Bug fixes
 - Fixed the problem when loading multiple structures with components.